requirements.txt
databricks-sdk
databricks-sql-connector
duckdb
dash
pandas
pyarrow
plotly